requirements.txt

flask
pypdf
python-pptx
python-docx
openpyxl
